matrix-transposition
Given a linked list and a value x, partition it such that all nodes less than x come before nodes greater than or equal to x. You should preserve the original relative order of the nodes in each of the two partitions.

For example, given the linked list 1 -> 4 -> 3 -> 2 -> 5 -> 2 and x = 3, your function should return the partitioned list as 1 -> 2 -> 2 -> 4 -> 3 -> 5.